+++ HealthTracker.API/SimpleData.cs
@@ -7,7 +7,7 @@ namespace HealthTracker.API
         public DeviceKind Kind { get; set; }
         public ulong Origin { get; set; }
         public DateTimeOffset Time { get; set; }
-        public float Data { get; set; }
+        public double Data { get; set; }
         public Unit Unit { get; set; }
         public DataKind Status { get; set; }
     }
